% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_custom.R
\name{lda_custom}
\alias{lda_custom}
\title{Custom LDA function}
\usage{
lda_custom(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\value{
a custom LDA function
}
\description{
Custom LDA function
}
\examples{
# put your awesome examples here
}
